+++ abi/bind/util.go
@@ -18,20 +18,59 @@ package bind
 
 import (
 	"context"
+	"errors"
+	"time"
 
-	"github.com/ethereum/go-ethereum/abi/bind"
 	"github.com/ethereum/go-ethereum/common"
 	"github.com/ethereum/go-ethereum/core/types"
+	"github.com/ethereum/go-ethereum/log"
 )
 
 // WaitMined waits for tx to be mined on the blockchain.
 // It stops waiting when the context is canceled.
 func WaitMined(ctx context.Context, b DeployBackend, tx *types.Transaction) (*types.Receipt, error) {
-	return bind.WaitMined(ctx, b, tx)
+	queryTicker := time.NewTicker(time.Second)
+	defer queryTicker.Stop()
+
+	logger := log.New("hash", tx.Hash())
+	for {
+		receipt, err := b.TransactionReceipt(ctx, tx.Hash())
+		if receipt != nil {
+			return receipt, nil
+		}
+		if err != nil {
+			logger.Trace("Receipt retrieval failed", "err", err)
+		} else {
+			logger.Trace("Transaction not yet mined")
+		}
+		// Wait for the next round.
+		select {
+		case <-ctx.Done():
+			return nil, ctx.Err()
+		case <-queryTicker.C:
+		}
+	}
 }
 
 // WaitDeployed waits for a contract deployment transaction and returns the on-chain
 // contract address when it is mined. It stops waiting when ctx is canceled.
 func WaitDeployed(ctx context.Context, b DeployBackend, tx *types.Transaction) (common.Address, error) {
-	return bind.WaitDeployed(ctx, b, tx)
+	if tx.To() != nil {
+		return common.Address{}, errors.New("tx is not contract creation")
+	}
+	receipt, err := WaitMined(ctx, b, tx)
+	if err != nil {
+		return common.Address{}, err
+	}
+	if receipt.ContractAddress == (common.Address{}) {
+		return common.Address{}, errors.New("zero address")
+	}
+	// Check that code has indeed been deployed at the address.
+	// This matters on pre-Homestead chains: OOG in the constructor
+	// could leave an empty account behind.
+	code, err := b.CodeAt(ctx, receipt.ContractAddress, nil)
+	if err == nil && len(code) == 0 {
+		err = ErrNoCodeAfterDeploy
+	}
+	return receipt.ContractAddress, err
 }
