+++ accounts/abi/abi.go
@@ -17,263 +17,25 @@
 package abi
 
 import (
-	"bytes"
-	"encoding/json"
-	"errors"
-	"fmt"
 	"io"
 
-	"github.com/ethereum/go-ethereum/common"
-	"github.com/ethereum/go-ethereum/crypto"
+	"github.com/ethereum/go-ethereum/abi"
 )
 
 // The ABI holds information about a contract's context and available
 // invokable methods. It will allow you to type check function calls and
 // packs data accordingly.
-type ABI struct {
-	Constructor Method
-	Methods     map[string]Method
-	Events      map[string]Event
-
-	// Additional "special" functions introduced in solidity v0.6.0.
-	// It's separated from the original default fallback. Each contract
-	// can only define one fallback and receive function.
-	Fallback Method // Note it's also used to represent legacy fallback before v0.6.0
-	Receive  Method
-}
+type ABI = abi.ABI
 
 // JSON returns a parsed ABI interface and error if it failed.
 func JSON(reader io.Reader) (ABI, error) {
-	dec := json.NewDecoder(reader)
-
-	var abi ABI
-	if err := dec.Decode(&abi); err != nil {
-		return ABI{}, err
-	}
-	return abi, nil
+	return abi.JSON(reader)
 }
-
-// Pack the given method name to conform the ABI. Method call's data
-// will consist of method_id, args0, arg1, ... argN. Method id consists
-// of 4 bytes and arguments are all 32 bytes.
-// Method ids are created from the first 4 bytes of the hash of the
-// methods string signature. (signature = baz(uint32,string32))
-func (abi ABI) Pack(name string, args ...interface{}) ([]byte, error) {
-	// Fetch the ABI of the requested method
-	if name == "" {
-		// constructor
-		arguments, err := abi.Constructor.Inputs.Pack(args...)
-		if err != nil {
-			return nil, err
-		}
-		return arguments, nil
-	}
-	method, exist := abi.Methods[name]
-	if !exist {
-		return nil, fmt.Errorf("method '%s' not found", name)
-	}
-	arguments, err := method.Inputs.Pack(args...)
-	if err != nil {
-		return nil, err
-	}
-	// Pack up the method ID too if not a constructor and return
-	return append(method.ID, arguments...), nil
-}
-
-func (abi ABI) getArguments(name string, data []byte) (Arguments, error) {
-	// since there can't be naming collisions with contracts and events,
-	// we need to decide whether we're calling a method or an event
-	var args Arguments
-	if method, ok := abi.Methods[name]; ok {
-		if len(data)%32 != 0 {
-			return nil, fmt.Errorf("abi: improperly formatted output: %s - Bytes: [%+v]", string(data), data)
-		}
-		args = method.Outputs
-	}
-	if event, ok := abi.Events[name]; ok {
-		args = event.Inputs
-	}
-	if args == nil {
-		return nil, errors.New("abi: could not locate named method or event")
-	}
-	return args, nil
-}
-
-// Unpack unpacks the output according to the abi specification.
-func (abi ABI) Unpack(name string, data []byte) ([]interface{}, error) {
-	args, err := abi.getArguments(name, data)
-	if err != nil {
-		return nil, err
-	}
-	return args.Unpack(data)
-}
-
-// UnpackIntoInterface unpacks the output in v according to the abi specification.
-// It performs an additional copy. Please only use, if you want to unpack into a
-// structure that does not strictly conform to the abi structure (e.g. has additional arguments)
-func (abi ABI) UnpackIntoInterface(v interface{}, name string, data []byte) error {
-	args, err := abi.getArguments(name, data)
-	if err != nil {
-		return err
-	}
-	unpacked, err := args.Unpack(data)
-	if err != nil {
-		return err
-	}
-	return args.Copy(v, unpacked)
-}
-
-// UnpackIntoMap unpacks a log into the provided map[string]interface{}.
-func (abi ABI) UnpackIntoMap(v map[string]interface{}, name string, data []byte) (err error) {
-	args, err := abi.getArguments(name, data)
-	if err != nil {
-		return err
-	}
-	return args.UnpackIntoMap(v, data)
-}
-
-// UnmarshalJSON implements json.Unmarshaler interface.
-func (abi *ABI) UnmarshalJSON(data []byte) error {
-	var fields []struct {
-		Type    string
-		Name    string
-		Inputs  []Argument
-		Outputs []Argument
-
-		// Status indicator which can be: "pure", "view",
-		// "nonpayable" or "payable".
-		StateMutability string
-
-		// Deprecated Status indicators, but removed in v0.6.0.
-		Constant bool // True if function is either pure or view
-		Payable  bool // True if function is payable
-
-		// Event relevant indicator represents the event is
-		// declared as anonymous.
-		Anonymous bool
-	}
-	if err := json.Unmarshal(data, &fields); err != nil {
-		return err
-	}
-	abi.Methods = make(map[string]Method)
-	abi.Events = make(map[string]Event)
-	for _, field := range fields {
-		switch field.Type {
-		case "constructor":
-			abi.Constructor = NewMethod("", "", Constructor, field.StateMutability, field.Constant, field.Payable, field.Inputs, nil)
-		case "function":
-			name := abi.overloadedMethodName(field.Name)
-			abi.Methods[name] = NewMethod(name, field.Name, Function, field.StateMutability, field.Constant, field.Payable, field.Inputs, field.Outputs)
-		case "fallback":
-			// New introduced function type in v0.6.0, check more detail
-			// here https://solidity.readthedocs.io/en/v0.6.0/contracts.html#fallback-function
-			if abi.HasFallback() {
-				return errors.New("only single fallback is allowed")
-			}
-			abi.Fallback = NewMethod("", "", Fallback, field.StateMutability, field.Constant, field.Payable, nil, nil)
-		case "receive":
-			// New introduced function type in v0.6.0, check more detail
-			// here https://solidity.readthedocs.io/en/v0.6.0/contracts.html#fallback-function
-			if abi.HasReceive() {
-				return errors.New("only single receive is allowed")
-			}
-			if field.StateMutability != "payable" {
-				return errors.New("the statemutability of receive can only be payable")
-			}
-			abi.Receive = NewMethod("", "", Receive, field.StateMutability, field.Constant, field.Payable, nil, nil)
-		case "event":
-			name := abi.overloadedEventName(field.Name)
-			abi.Events[name] = NewEvent(name, field.Name, field.Anonymous, field.Inputs)
-		default:
-			return fmt.Errorf("abi: could not recognize type %v of field %v", field.Type, field.Name)
-		}
-	}
-	return nil
-}
-
-// overloadedMethodName returns the next available name for a given function.
-// Needed since solidity allows for function overload.
-//
-// e.g. if the abi contains Methods send, send1
-// overloadedMethodName would return send2 for input send.
-func (abi *ABI) overloadedMethodName(rawName string) string {
-	name := rawName
-	_, ok := abi.Methods[name]
-	for idx := 0; ok; idx++ {
-		name = fmt.Sprintf("%s%d", rawName, idx)
-		_, ok = abi.Methods[name]
-	}
-	return name
-}
-
-// overloadedEventName returns the next available name for a given event.
-// Needed since solidity allows for event overload.
-//
-// e.g. if the abi contains events received, received1
-// overloadedEventName would return received2 for input received.
-func (abi *ABI) overloadedEventName(rawName string) string {
-	name := rawName
-	_, ok := abi.Events[name]
-	for idx := 0; ok; idx++ {
-		name = fmt.Sprintf("%s%d", rawName, idx)
-		_, ok = abi.Events[name]
-	}
-	return name
-}
-
-// MethodById looks up a method by the 4-byte id,
-// returns nil if none found.
-func (abi *ABI) MethodById(sigdata []byte) (*Method, error) {
-	if len(sigdata) < 4 {
-		return nil, fmt.Errorf("data too short (%d bytes) for abi method lookup", len(sigdata))
-	}
-	for _, method := range abi.Methods {
-		if bytes.Equal(method.ID, sigdata[:4]) {
-			return &method, nil
-		}
-	}
-	return nil, fmt.Errorf("no method with id: %#x", sigdata[:4])
-}
-
-// EventByID looks an event up by its topic hash in the
-// ABI and returns nil if none found.
-func (abi *ABI) EventByID(topic common.Hash) (*Event, error) {
-	for _, event := range abi.Events {
-		if bytes.Equal(event.ID.Bytes(), topic.Bytes()) {
-			return &event, nil
-		}
-	}
-	return nil, fmt.Errorf("no event with id: %#x", topic.Hex())
-}
-
-// HasFallback returns an indicator whether a fallback function is included.
-func (abi *ABI) HasFallback() bool {
-	return abi.Fallback.Type == Fallback
-}
-
-// HasReceive returns an indicator whether a receive function is included.
-func (abi *ABI) HasReceive() bool {
-	return abi.Receive.Type == Receive
-}
-
-// revertSelector is a special function selector for revert reason unpacking.
-var revertSelector = crypto.Keccak256([]byte("Error(string)"))[:4]
 
 // UnpackRevert resolves the abi-encoded revert reason. According to the solidity
 // spec https://solidity.readthedocs.io/en/latest/control-structures.html#revert,
 // the provided revert reason is abi-encoded as if it were a call to a function
 // `Error(string)`. So it's a special tool for it.
 func UnpackRevert(data []byte) (string, error) {
-	if len(data) < 4 {
-		return "", errors.New("invalid data for unpacking")
-	}
-	if !bytes.Equal(data[:4], revertSelector) {
-		return "", errors.New("invalid data for unpacking")
-	}
-	typ, _ := NewType("string", "", nil)
-	unpacked, err := (Arguments{{Type: typ}}).Unpack(data[4:])
-	if err != nil {
-		return "", err
-	}
-	return unpacked[0].(string), nil
+	return abi.UnpackRevert(data)
 }
