+++ accounts/abi/bind/backend.go
@@ -17,96 +17,45 @@
 package bind
 
 import (
-	"context"
-	"errors"
-	"math/big"
-
-	"github.com/ethereum/go-ethereum"
-	"github.com/ethereum/go-ethereum/common"
-	"github.com/ethereum/go-ethereum/core/types"
+	"github.com/ethereum/go-ethereum/abi/bind"
 )
 
 var (
 	// ErrNoCode is returned by call and transact operations for which the requested
 	// recipient contract to operate on does not exist in the state db or does not
 	// have any code associated with it (i.e. suicided).
-	ErrNoCode = errors.New("no contract code at given address")
+	ErrNoCode = bind.ErrNoCode
 
 	// This error is raised when attempting to perform a pending state action
 	// on a backend that doesn't implement PendingContractCaller.
-	ErrNoPendingState = errors.New("backend does not support pending state")
+	ErrNoPendingState = bind.ErrNoPendingState
 
 	// This error is returned by WaitDeployed if contract creation leaves an
 	// empty contract behind.
-	ErrNoCodeAfterDeploy = errors.New("no contract code after deployment")
+	ErrNoCodeAfterDeploy = bind.ErrNoCodeAfterDeploy
 )
 
 // ContractCaller defines the methods needed to allow operating with a contract on a read
 // only basis.
-type ContractCaller interface {
-	// CodeAt returns the code of the given account. This is needed to differentiate
-	// between contract internal errors and the local chain being out of sync.
-	CodeAt(ctx context.Context, contract common.Address, blockNumber *big.Int) ([]byte, error)
-	// ContractCall executes an Ethereum contract call with the specified data as the
-	// input.
-	CallContract(ctx context.Context, call ethereum.CallMsg, blockNumber *big.Int) ([]byte, error)
-}
+type ContractCaller = bind.ContractCaller
 
 // PendingContractCaller defines methods to perform contract calls on the pending state.
 // Call will try to discover this interface when access to the pending state is requested.
 // If the backend does not support the pending state, Call returns ErrNoPendingState.
-type PendingContractCaller interface {
-	// PendingCodeAt returns the code of the given account in the pending state.
-	PendingCodeAt(ctx context.Context, contract common.Address) ([]byte, error)
-	// PendingCallContract executes an Ethereum contract call against the pending state.
-	PendingCallContract(ctx context.Context, call ethereum.CallMsg) ([]byte, error)
-}
+type PendingContractCaller = bind.PendingContractCaller
 
 // ContractTransactor defines the methods needed to allow operating with a contract
 // on a write only basis. Besides the transacting method, the remainder are helpers
 // used when the user does not provide some needed values, but rather leaves it up
 // to the transactor to decide.
-type ContractTransactor interface {
-	// PendingCodeAt returns the code of the given account in the pending state.
-	PendingCodeAt(ctx context.Context, account common.Address) ([]byte, error)
-	// PendingNonceAt retrieves the current pending nonce associated with an account.
-	PendingNonceAt(ctx context.Context, account common.Address) (uint64, error)
-	// SuggestGasPrice retrieves the currently suggested gas price to allow a timely
-	// execution of a transaction.
-	SuggestGasPrice(ctx context.Context) (*big.Int, error)
-	// EstimateGas tries to estimate the gas needed to execute a specific
-	// transaction based on the current pending state of the backend blockchain.
-	// There is no guarantee that this is the true gas limit requirement as other
-	// transactions may be added or removed by miners, but it should provide a basis
-	// for setting a reasonable default.
-	EstimateGas(ctx context.Context, call ethereum.CallMsg) (gas uint64, err error)
-	// SendTransaction injects the transaction into the pending pool for execution.
-	SendTransaction(ctx context.Context, tx *types.Transaction) error
-}
+type ContractTransactor = bind.ContractTransactor
 
 // ContractFilterer defines the methods needed to access log events using one-off
 // queries or continuous event subscriptions.
-type ContractFilterer interface {
-	// FilterLogs executes a log filter operation, blocking during execution and
-	// returning all the results in one batch.
-	//
-	// TODO(karalabe): Deprecate when the subscription one can return past data too.
-	FilterLogs(ctx context.Context, query ethereum.FilterQuery) ([]types.Log, error)
-
-	// SubscribeFilterLogs creates a background log filtering operation, returning
-	// a subscription immediately, which can be used to stream the found events.
-	SubscribeFilterLogs(ctx context.Context, query ethereum.FilterQuery, ch chan<- types.Log) (ethereum.Subscription, error)
-}
+type ContractFilterer = bind.ContractFilterer
 
 // DeployBackend wraps the operations needed by WaitMined and WaitDeployed.
-type DeployBackend interface {
-	TransactionReceipt(ctx context.Context, txHash common.Hash) (*types.Receipt, error)
-	CodeAt(ctx context.Context, account common.Address, blockNumber *big.Int) ([]byte, error)
-}
+type DeployBackend = bind.DeployBackend
 
 // ContractBackend defines the methods needed to work with contracts on a read-write basis.
-type ContractBackend interface {
-	ContractCaller
-	ContractTransactor
-	ContractFilterer
-}
+type ContractBackend = bind.ContractBackend
