+++ accounts/abi/unpack.go
@@ -17,284 +17,22 @@
 package abi
 
 import (
-	"encoding/binary"
-	"fmt"
-	"math/big"
-	"reflect"
-
-	"github.com/ethereum/go-ethereum/common"
+	"github.com/ethereum/go-ethereum/abi"
 )
 
 var (
 	// MaxUint256 is the maximum value that can be represented by a uint256.
-	MaxUint256 = new(big.Int).Sub(new(big.Int).Lsh(common.Big1, 256), common.Big1)
+	MaxUint256 = abi.MaxUint256
 	// MaxInt256 is the maximum value that can be represented by a int256.
-	MaxInt256 = new(big.Int).Sub(new(big.Int).Lsh(common.Big1, 255), common.Big1)
+	MaxInt256 = abi.MaxInt256
 )
 
 // ReadInteger reads the integer based on its kind and returns the appropriate value.
 func ReadInteger(typ Type, b []byte) interface{} {
-	if typ.T == UintTy {
-		switch typ.Size {
-		case 8:
-			return b[len(b)-1]
-		case 16:
-			return binary.BigEndian.Uint16(b[len(b)-2:])
-		case 32:
-			return binary.BigEndian.Uint32(b[len(b)-4:])
-		case 64:
-			return binary.BigEndian.Uint64(b[len(b)-8:])
-		default:
-			// the only case left for unsigned integer is uint256.
-			return new(big.Int).SetBytes(b)
-		}
-	}
-	switch typ.Size {
-	case 8:
-		return int8(b[len(b)-1])
-	case 16:
-		return int16(binary.BigEndian.Uint16(b[len(b)-2:]))
-	case 32:
-		return int32(binary.BigEndian.Uint32(b[len(b)-4:]))
-	case 64:
-		return int64(binary.BigEndian.Uint64(b[len(b)-8:]))
-	default:
-		// the only case left for integer is int256
-		// big.SetBytes can't tell if a number is negative or positive in itself.
-		// On EVM, if the returned number > max int256, it is negative.
-		// A number is > max int256 if the bit at position 255 is set.
-		ret := new(big.Int).SetBytes(b)
-		if ret.Bit(255) == 1 {
-			ret.Add(MaxUint256, new(big.Int).Neg(ret))
-			ret.Add(ret, common.Big1)
-			ret.Neg(ret)
-		}
-		return ret
-	}
-}
-
-// readBool reads a bool.
-func readBool(word []byte) (bool, error) {
-	for _, b := range word[:31] {
-		if b != 0 {
-			return false, errBadBool
-		}
-	}
-	switch word[31] {
-	case 0:
-		return false, nil
-	case 1:
-		return true, nil
-	default:
-		return false, errBadBool
-	}
-}
-
-// A function type is simply the address with the function selection signature at the end.
-//
-// readFunctionType enforces that standard by always presenting it as a 24-array (address + sig = 24 bytes)
-func readFunctionType(t Type, word []byte) (funcTy [24]byte, err error) {
-	if t.T != FunctionTy {
-		return [24]byte{}, fmt.Errorf("abi: invalid type in call to make function type byte array")
-	}
-	if garbage := binary.BigEndian.Uint64(word[24:32]); garbage != 0 {
-		err = fmt.Errorf("abi: got improperly encoded function type, got %v", word)
-	} else {
-		copy(funcTy[:], word[0:24])
-	}
-	return
+	return abi.ReadInteger(typ, b)
 }
 
 // ReadFixedBytes uses reflection to create a fixed array to be read from.
 func ReadFixedBytes(t Type, word []byte) (interface{}, error) {
-	if t.T != FixedBytesTy {
-		return nil, fmt.Errorf("abi: invalid type in call to make fixed byte array")
-	}
-	// convert
-	array := reflect.New(t.GetType()).Elem()
-
-	reflect.Copy(array, reflect.ValueOf(word[0:t.Size]))
-	return array.Interface(), nil
-
-}
-
-// forEachUnpack iteratively unpack elements.
-func forEachUnpack(t Type, output []byte, start, size int) (interface{}, error) {
-	if size < 0 {
-		return nil, fmt.Errorf("cannot marshal input to array, size is negative (%d)", size)
-	}
-	if start+32*size > len(output) {
-		return nil, fmt.Errorf("abi: cannot marshal in to go array: offset %d would go over slice boundary (len=%d)", len(output), start+32*size)
-	}
-
-	// this value will become our slice or our array, depending on the type
-	var refSlice reflect.Value
-
-	if t.T == SliceTy {
-		// declare our slice
-		refSlice = reflect.MakeSlice(t.GetType(), size, size)
-	} else if t.T == ArrayTy {
-		// declare our array
-		refSlice = reflect.New(t.GetType()).Elem()
-	} else {
-		return nil, fmt.Errorf("abi: invalid type in array/slice unpacking stage")
-	}
-
-	// Arrays have packed elements, resulting in longer unpack steps.
-	// Slices have just 32 bytes per element (pointing to the contents).
-	elemSize := getTypeSize(*t.Elem)
-
-	for i, j := start, 0; j < size; i, j = i+elemSize, j+1 {
-		inter, err := toGoType(i, *t.Elem, output)
-		if err != nil {
-			return nil, err
-		}
-
-		// append the item to our reflect slice
-		refSlice.Index(j).Set(reflect.ValueOf(inter))
-	}
-
-	// return the interface
-	return refSlice.Interface(), nil
-}
-
-func forTupleUnpack(t Type, output []byte) (interface{}, error) {
-	retval := reflect.New(t.GetType()).Elem()
-	virtualArgs := 0
-	for index, elem := range t.TupleElems {
-		marshalledValue, err := toGoType((index+virtualArgs)*32, *elem, output)
-		if elem.T == ArrayTy && !isDynamicType(*elem) {
-			// If we have a static array, like [3]uint256, these are coded as
-			// just like uint256,uint256,uint256.
-			// This means that we need to add two 'virtual' arguments when
-			// we count the index from now on.
-			//
-			// Array values nested multiple levels deep are also encoded inline:
-			// [2][3]uint256: uint256,uint256,uint256,uint256,uint256,uint256
-			//
-			// Calculate the full array size to get the correct offset for the next argument.
-			// Decrement it by 1, as the normal index increment is still applied.
-			virtualArgs += getTypeSize(*elem)/32 - 1
-		} else if elem.T == TupleTy && !isDynamicType(*elem) {
-			// If we have a static tuple, like (uint256, bool, uint256), these are
-			// coded as just like uint256,bool,uint256
-			virtualArgs += getTypeSize(*elem)/32 - 1
-		}
-		if err != nil {
-			return nil, err
-		}
-		retval.Field(index).Set(reflect.ValueOf(marshalledValue))
-	}
-	return retval.Interface(), nil
-}
-
-// toGoType parses the output bytes and recursively assigns the value of these bytes
-// into a go type with accordance with the ABI spec.
-func toGoType(index int, t Type, output []byte) (interface{}, error) {
-	if index+32 > len(output) {
-		return nil, fmt.Errorf("abi: cannot marshal in to go type: length insufficient %d require %d", len(output), index+32)
-	}
-
-	var (
-		returnOutput  []byte
-		begin, length int
-		err           error
-	)
-
-	// if we require a length prefix, find the beginning word and size returned.
-	if t.requiresLengthPrefix() {
-		begin, length, err = lengthPrefixPointsTo(index, output)
-		if err != nil {
-			return nil, err
-		}
-	} else {
-		returnOutput = output[index : index+32]
-	}
-
-	switch t.T {
-	case TupleTy:
-		if isDynamicType(t) {
-			begin, err := tuplePointsTo(index, output)
-			if err != nil {
-				return nil, err
-			}
-			return forTupleUnpack(t, output[begin:])
-		}
-		return forTupleUnpack(t, output[index:])
-	case SliceTy:
-		return forEachUnpack(t, output[begin:], 0, length)
-	case ArrayTy:
-		if isDynamicType(*t.Elem) {
-			offset := binary.BigEndian.Uint64(returnOutput[len(returnOutput)-8:])
-			if offset > uint64(len(output)) {
-				return nil, fmt.Errorf("abi: toGoType offset greater than output length: offset: %d, len(output): %d", offset, len(output))
-			}
-			return forEachUnpack(t, output[offset:], 0, t.Size)
-		}
-		return forEachUnpack(t, output[index:], 0, t.Size)
-	case StringTy: // variable arrays are written at the end of the return bytes
-		return string(output[begin : begin+length]), nil
-	case IntTy, UintTy:
-		return ReadInteger(t, returnOutput), nil
-	case BoolTy:
-		return readBool(returnOutput)
-	case AddressTy:
-		return common.BytesToAddress(returnOutput), nil
-	case HashTy:
-		return common.BytesToHash(returnOutput), nil
-	case BytesTy:
-		return output[begin : begin+length], nil
-	case FixedBytesTy:
-		return ReadFixedBytes(t, returnOutput)
-	case FunctionTy:
-		return readFunctionType(t, returnOutput)
-	default:
-		return nil, fmt.Errorf("abi: unknown type %v", t.T)
-	}
-}
-
-// lengthPrefixPointsTo interprets a 32 byte slice as an offset and then determines which indices to look to decode the type.
-func lengthPrefixPointsTo(index int, output []byte) (start int, length int, err error) {
-	bigOffsetEnd := big.NewInt(0).SetBytes(output[index : index+32])
-	bigOffsetEnd.Add(bigOffsetEnd, common.Big32)
-	outputLength := big.NewInt(int64(len(output)))
-
-	if bigOffsetEnd.Cmp(outputLength) > 0 {
-		return 0, 0, fmt.Errorf("abi: cannot marshal in to go slice: offset %v would go over slice boundary (len=%v)", bigOffsetEnd, outputLength)
-	}
-
-	if bigOffsetEnd.BitLen() > 63 {
-		return 0, 0, fmt.Errorf("abi offset larger than int64: %v", bigOffsetEnd)
-	}
-
-	offsetEnd := int(bigOffsetEnd.Uint64())
-	lengthBig := big.NewInt(0).SetBytes(output[offsetEnd-32 : offsetEnd])
-
-	totalSize := big.NewInt(0)
-	totalSize.Add(totalSize, bigOffsetEnd)
-	totalSize.Add(totalSize, lengthBig)
-	if totalSize.BitLen() > 63 {
-		return 0, 0, fmt.Errorf("abi: length larger than int64: %v", totalSize)
-	}
-
-	if totalSize.Cmp(outputLength) > 0 {
-		return 0, 0, fmt.Errorf("abi: cannot marshal in to go type: length insufficient %v require %v", outputLength, totalSize)
-	}
-	start = int(bigOffsetEnd.Uint64())
-	length = int(lengthBig.Uint64())
-	return
-}
-
-// tuplePointsTo resolves the location reference for dynamic tuple.
-func tuplePointsTo(index int, output []byte) (start int, err error) {
-	offset := big.NewInt(0).SetBytes(output[index : index+32])
-	outputLen := big.NewInt(int64(len(output)))
-
-	if offset.Cmp(big.NewInt(int64(len(output)))) > 0 {
-		return 0, fmt.Errorf("abi: cannot marshal in to go slice: offset %v would go over slice boundary (len=%v)", offset, outputLen)
-	}
-	if offset.BitLen() > 63 {
-		return 0, fmt.Errorf("abi offset larger than int64: %v", offset)
-	}
-	return int(offset.Uint64()), nil
+	return abi.ReadFixedBytes(t, word)
 }
